mcp_agent_app/requirements.txt
mcp_agent[google]==0.2.5
aiohttp
pymongo
python-dotenv